manager: drop remaining WAL in case of error
The error-management story is still somewhat wrong though since in
principle, a batch of entries in the WAL should be treated as a
transaction which must complete atomically. We don't support rolling
back operations, so we're only halfway there. Nonetheless, this should
not put the system into undefined state: a partially completed
transaction in our case still leaves the system in a state where
things sort of work. The worst that could happen is a BootWorker
failing, thus preventing a ProxyRefresh from occurring. This could
leave some number of workers running, but inaccessible.